Move EffectsComposer into TS
